*Setting page에서 자동 로그인 해제시 다시 로그인 페이지로 이동하게 수정

+++ GoodBetter/www/js/Setting.js
@@ -2,6 +2,13 @@
  * Created by minseock on 2015-07-08.
  */
 function disableAutoLogin() {
-    window.localStorage.setItem("auto",false);
-    alert("자동 로그인이 해제되었습니다.");
+	var r = confirm("자동 로그인을 해제하시겠습니까?");
+	if (r == true){
+		window.localStorage.setItem("auto",false);
+		alert("자동 로그인이 해제되었습니다.");
+		
+		var dirPath = dirname(location.href);
+		fullPath = dirPath + "/index.html";
+		window.location = fullPath;
+	}
 }
